requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
black==24.10.0
boto3==1.35.39
botocore==1.35.39
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.0
click==8.1.7
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
deepl==1.19.1
defusedxml==0.8.0rc2
dj-rest-auth==6.0.0
Django==5.1.2
django-autoslug==1.9.9
django-cleanup==9.0.0
django-cors-headers==4.5.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-modeltranslation==0.19.9
django-redis==5.4.0
django-storages==1.14.4
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
environs==11.0.0
fcm-django==2.2.1
fernet==1.0.1
firebase-admin==6.5.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
isort==5.13.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
marshmallow==3.22.0
msgpack==1.1.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.9
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyotp==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
redis==5.1.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.3
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
swapper==1.4.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.7.0
zope.interface==7.1.0
